requirements.txt
scrapli[asyncssh]
python-dotenv